+++ backend/canisters/storage_bucket/impl/src/lifecycle/mod.rs
@@ -10,8 +10,6 @@ mod init;
 mod post_upgrade;
 mod pre_upgrade;
 
-const BUFFER_SIZE: usize = 16 * 1024 * 1024; // 16MB
-
 fn init_env() -> Box<CanisterEnv> {
     ic_cdk_timers::set_timer(Duration::ZERO, reseed_rng);
     Box::default()