+++ backend/libraries/types/src/lib.rs
@@ -1,7 +1,7 @@
 use candid::{CandidType, Principal};
 use ic_ledger_types::Tokens;
 use serde::{Deserialize, Serialize};
-use std::collections::{BTreeMap, HashSet};
+use std::collections::{BTreeMap, HashMap, HashSet};
 
 mod avatar;
 mod bots;
@@ -157,6 +157,10 @@ impl<T: PartialEq> PushIfNotContains<T> for Vec<T> {
 }
 
 pub fn is_empty_slice<T>(value: &[T]) -> bool {
+    value.is_empty()
+}
+
+pub fn is_empty_hashmap<K, V>(value: &HashMap<K, V>) -> bool {
     value.is_empty()
 }